requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
cryptography==46.0.3
datapizza-ai==0.0.9
datapizza-ai-clients-openai==0.0.9
datapizza-ai-core==0.0.15
datapizza-ai-embedders-openai==0.0.5
datapizza-ai-parsers-docling==0.0.8
datapizza-ai-vectorstores-qdrant==0.0.6
dill==0.4.0
distlib==0.4.0
distro==1.9.0
docling==2.64.1
docling-core==2.55.0
docling-ibm-models==3.10.3
docling-parse==4.7.2
easyocr==1.7.2
et-xmlfile==2.0.0
faker==38.2.0
filelock==3.20.0
filetype==1.2.0
fsspec==2025.12.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
imageio==2.37.2
importlib-metadata==8.7.0
iniconfig==2.3.0
jinja2==3.1.6
jiter==0.12.0
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
lazy-loader==0.4
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.1
markupsafe==3.0.3
mcp==1.23.1
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.6.1
ninja==1.13.0
nodeenv==1.9.1
numpy==2.3.5
ocrmac==1.0.0
omegaconf==2.3.0
openai==2.9.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pip @ file:///Applications/PyCharm.app/Contents/plugins/python-ce/helpers/pip-24.3.1-py2.py3-none-any.whl
platformdirs==4.5.1
pluggy==1.6.0
polyfactory==3.1.0
portalocker==3.2.0
pre-commit==4.5.0
protobuf==6.33.1
psutil==7.1.3
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pylatexenc==2.10
pyobjc-core==12.1
pyobjc-framework-cocoa==12.1
pyobjc-framework-coreml==12.1
pyobjc-framework-quartz==12.1
pyobjc-framework-vision==12.1
pypdfium2==4.30.0
pytest==9.0.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.3
qdrant-client==1.16.1
rapidocr==3.4.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rtree==1.4.1
ruff==0.14.8
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.16.3
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.12.12
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.0
uvicorn==0.38.0
virtualenv==20.35.4
xlsxwriter==3.2.9
zipp==3.23.0
